Add a LICENSE.md
The license was described in package, but the MIT license text says that
a copy of it should be included in the code base.  This will make it
easier for potential users / contributors to know for sure that the
license applies to the entire code base.

Issue #219 Consider adding a LICENSE.md